requirements.txt
aiofiles==24.1.0
aiogram==3.17.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
asyncpg==0.30.0
attrs==24.3.0
black==24.10.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
exceptiongroup==1.2.2
flake8==7.1.1
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
magic-filter==1.0.12
Mako==1.3.9
MarkupSafe==3.0.2
mccabe==0.7.0
multidict==6.1.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.6
pluggy==1.5.0
propcache==0.2.1
psycopg2-binary==2.9.10
pycodestyle==2.12.1
pydantic==2.10.5
pydantic_core==2.27.2
pyflakes==3.2.0
pyTelegramBotAPI==4.26.0
pytest==8.3.5
python-decouple==3.8
python-dotenv==1.0.1
python-telegram-bot==21.10
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.38
tomli==2.2.1
typing_extensions==4.12.2
urllib3==2.3.0
yarl==1.18.3
